Added SSO
Added in a Single Sign On Feature that allows the user to sign in using their Google Account rather than typing in their email and password. It uses Firebase Auth to connect to the google services to verify the user's account before allowing the user to sign in using google.